// frontend/frontend/src/app/models/profile.ts
export class profile {
    id?:string;
    name?: string;
    email?: string;
    mobile?: number;
    address1?: string;
    address2?: string;
    postcode?: string;
    state?: string;
   mobile1?:string;
    age?:number;
    country?: string;
    imageUrl?: string;
   
  }


   // userProfile?: profile[]; 
    // image?: File; 